body-table-container become component
