+++ addon/views/header-table-container.js
@@ -1,9 +1,9 @@
 import Ember from 'ember';
-import TableContainer from 'ember-table/views/table-container';
+import TableContainer from 'ember-table/mixins/table-container';
 import ShowHorizontalScrollMixin from 'ember-table/mixins/show-horizontal-scroll';
 import RegisterTableComponentMixin from 'ember-table/mixins/register-table-component';
 
-export default TableContainer.extend(
+export default Ember.View.extend( TableContainer,
 ShowHorizontalScrollMixin, RegisterTableComponentMixin, {
   templateName: 'header-table-container',
   classNames: ['ember-table-table-container',
