+++ addon/views/lazy-table-block.js
@@ -6,12 +6,8 @@ export default LazyContainerView.extend(
 RegisterTableComponentMixin, {
   classNames: ['ember-table-table-block'],
   styleBindings: ['width'],
-  itemViewClass: Ember.computed.alias('tableComponent.tableRowViewClass'),
-  rowHeight: Ember.computed.alias('tableComponent.rowHeight'),
   columns: null,
-  content: null,
   scrollLeft: null,
-  scrollTop: null,
 
   onScrollLeftDidChange: Ember.observer('scrollLeft', function() {
     this.$().scrollLeft(this.get('scrollLeft'));
