+++ app/components/body-table-container.js
@@ -1,0 +1,3 @@
+import BodyTableContainer from 'ember-table/components/body-table-container';
+
+export default BodyTableContainer;
