+++ app/templates/components/body-table-container.hbs
@@ -5,10 +5,12 @@
         {{view "lazy-table-block"
           content=bodyContent
           columns=fixedColumns
-          width=_fixedBlockWidth
+          width=fixedBlockWidth
           numItemsShowing=_numItemsShowing
           scrollTop=_scrollTop
           startIndex=_startIndex
+          itemViewClass=itemViewClass
+          rowHeight=rowHeight
           class="ember-table-left-table-block"
         }}
       {{/if}}
@@ -16,10 +18,12 @@
         content=bodyContent
         columns=tableColumns
         scrollLeft=_tableScrollLeft
-        width=_tableBlockWidth
+        width=tableBlockWidth
         numItemsShowing=_numItemsShowing
         scrollTop=_scrollTop
         startIndex=_startIndex
+        itemViewClass=itemViewClass
+        rowHeight=rowHeight
         class="ember-table-right-table-block"
       }}
     </div>
