--- app/views/body-table-container.js
@@ -1,3 +1,0 @@
-import BodyTableContainer from 'ember-table/views/body-table-container';
-
-export default BodyTableContainer;
